Fix wget-nossl variant by disabling ntlm support (bug 15718 and 15731)
Fix wget-nossl variant by disabling ntlm support in wget-nossl.

https://dev.openwrt.org/ticket/15718#comment:13
https://dev.openwrt.org/ticket/15731

In wget 1.15 the ntlm support requires either ssl or libnettle library.
Check-in behind the change:
http://git.savannah.gnu.org/cgit/wget.git/commit/?id=c19d76c02483f070beb688d6fe6f5fafb5674a08

ssl is not present in the nossl variant and developers prefer to keep the
nossl variant small, so the natural alternative is to disable ntlm support in
the nossl variant, as a configure option for that purpose exists and a simple
change in Openwrt Makefile is enough.

Tested with ar71xx/wndr3700.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>


git-svn-id: svn://svn.openwrt.org/openwrt/packages@40595 3c298f89-4303-0410-b956-a3cf2f4a3e73